+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,16 +15,41 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
+// Home Page
 Route::get('/', [ListingController::class, 'index']);
 
-Route::get('/listings/create', [ListingController::class, 'create']);
+// Manage Listing Page
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
-Route::post('/listings', [ListingController::class, 'store']);
+// Create Listing Page
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+// Edit Listing Page
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
+// Show Listing Page
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+// Register Page
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+// Login Page
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Create Listing Endpoint
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
+
+// Update Listing Endpoint
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
+
+// Delete Listing Endpoint
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+// Create New User Endpoint
+Route::post('/users', [UserController::class, 'store']);
+
+// Logout Endpoint
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Authenticate User Endpoint
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
